requirements.txt
Flask
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy
marshmallow
flask_marshmallow
marshmallow-sqlalchemy
requests
python-dotenv